source_code/containers/python_fastapi/requirements.txt
fastapi
uvicorn
httpx
psycopg2-binary

pylint
pytest
mypy

black
coverage